tx late test, speed up siggen, late report, lml mux
There is a dedicated test to experiment with late tx packets

The siggen app uses numpy to generate the sinewave,
it seems to handle up to 6 Msps.

Fixed the late reporting and detection code

The lml mux was reversed for tx and it turned out to be the same as rx.