lml and rate adjustments
* Pass sample rate to UpdateExternalDataRate().
The API is expecting the sample rate, not mclk rate.
The timestamp logic also needs this to be correct.
We can change the call to take the mclk rate,
but it just all needs to be consistent,
PLL and FPGA tick rate.

* Move lml setup code to STREAM board because its device specific

* Swap TX IQ in the ConfigureLML_BB2RF()
Best I can tell, this is the correct muxing after power up.
I choose to swap it in the lms7002m API call because
it may refer to a swap in the lms7 and not in the FPGA.