Revert change to addr < 0x0100 cache mirroring
The reason to write channel B registers when the address is less than 0x0100 is that the values should be mirrored on each A+B cache below 0x0100 to mimic the behaviour of the chip reguardless of how MAC is set. Without this change, we must be very careful to set the mac to ChA before every call to registers below 0x0100 so the cached readback is correct.