Fix error reporting 0 code on error
This broke error reporting b5a7c8a0a2fef721b2aa03fd4cb425d1aa8faded
because the return code was not meant to be the same as the error number.
A lot of failed calls were returning 0 which means success.
I can see how returning the error number might be useful,
so I changed the default error number when not specified to be -1