LimeSuiteConfig and LIME_DLL define
Minor fix from previous commit for supporting dll and static:
BUILD_SHARED_LIBS will not be set by CMakeLists,
so its going to be up to client builds against libLimeSuite
to define LIME_DLL when a shared library is used.
The CMakeLists.txt sets LIME_DLL when BUILD_SHARED_LIBS
is set and that is working just fine.