Properly free libusb device lists
Not sure why the libusb_free_device_list call would only run if no
devices are found at all (which never happens on machines with USB
controllers because of root hubs). No references are held to devices
from the list, so it is safe to free the list and unref its devices
after finishing the loop.

This avoids a device leak list when probing for devices with SoapySDR,
which occured even when no LimeSDR devices were present.