limesdr - remove interface claim from enumeration
The basic goal here was to remove the cannot claim error print
when discovering devices that are already claimed by another process.

It turns out that we should not need to libusb_claim_interface() nor
call libusb_kernel_driver_active() in order to read the descriptors
which are used for filling in the connection handles.

These calls are also actually applied in the IConnection STREAM
when the actual device is opened for control and streaming endpoints.