As of commit 990ebebe there was no support in the library at all for the END_BURST features in SoapyLMS7 readStreamStatus.  This made it very difficult to manage transmit streams, as one could never know when the last of a transmit burst had been committed.
This version detects END_BURST at the start of the last transmission.

Here are the major (non whitespace) changes:

1. Add END_OF_BURST flag to lime::IStreamChannel::Metadata enum.
2. Fix doxygen comment in ConnectionSTREAM::WaitForSending to indicate
that the incoming buffer is formatted for the FPGA data stream.
3. Added () to complex while condition in WaitForSending.  Indented to make
it readable.
4. Protected return 0; in else clause of FinishDataSending as the conditional
compilation and such can really wreak havoc in code like this. (Hard to tell
where one clause ends and another begins...)
5. Added flags member and initializer to ConnectionSTREAM.h : USBTransferContext. This is to allow us to mark a buffer as END_BURST and have it pass through
the processing chain all the way through TransmitPacketsLoop and back to
ReadStreamStatus.
6. Mitigated, but did not fix "magic" number 4096 in ConnectionSTREAM::TransmitPacketsLoop.
7. Added logic to detect "rising edge" of END_BURST signal.
8. Added flag setting in ILimeSDRStreaming::WriteStream to mark a buffer
as the end of a burst.
9. ILimeSDRStreaming::ReadStreamStatus now checks the sawEndOfBurst status
in the channel's streamer, and passes it back in the metadata argument.