Warning fixes (#126)
* Fixed some typos

* Got a start on fixing all the warnings.

* in src/ConnectionSTREAM/ConnectionSTREAM.cpp, the code looked like this:

    if (WaitForSending(context, timeout)==false);
         AbortSending(ep);

---------------------------------

This was flagged by a compiler warning about "misleading indentation"
So I fixed it.

    if (WaitForSending(context, timeout)==false) {
         AbortSending(ep);
    }

Now it probably also does something like the original intent.  As it was,
SendData would

	 BeginDataSending
	 WaitForSending
	 AbortSending
	 return FinishDataSending.

Was that the original intent?

* More work on fixing warnings.

* code in TransmitPacketsLoop had this:

            unsigned bytesSent;
            if (this->WaitForSending(handles[bi], 1000) == true)
                bytesSent = this->FinishDataSending(&buffers[bi*bufferSize], bufferSize, handles[bi]);
            if (bytesSent != bufferSize)
            {
                for (auto value : stream->mTxStreams)
                    value->overflow++;
            }
            else
                totalBytesSent += bytesSent;
            bufferUsed[bi] = false;

The "bytesSent" value is undetermined if WaitForSending returns false.
But bytesSent will either be equal to bufferSize (and bump overflow counters)
or will goose the totalBytesSent count by some unknown amount.

This was caught by a compiler warning about an uninitialized variable.

* Fixes to support generation of Doxygen output.

No functional changes, all are in doxygen comments.  Most relate to
missing parameter descriptions.

* Update LMS7002M.cpp

Original only had bad indentation. Endl should be printed every iteration

* Update lms7002m_calibrations.c

Dummy() is used in Delay() when __cplusplus is not defined