SoapyLMS7 caches stream enable and disables on close
To avoid redundant enabling/disabling of the stream when using
activateStream for burst controls, a cache of the state is added to the
stream. The actual stream enable/disable is only performed when the
actual state differs. Further, we automatically disable the stream
(which stops the threads) when closeStream is called when the streams
were left enabled. This prevents some segfaults on cleanup.