cmake - reduce simd flags when packaging for x86
Fix an issue when packaging LimeSuite for x86
because the -march=native cant build portable code.

X86 machines come with varying SIMD support (AVX, SSE, etc)
This change defaults the SIMD support for x86 to SSE3
only when packaging builds (building into /usr is the hint).

Assuming that SSE3 is a safe assumption that most machines will have.

This change should not affect existing source builds in terms of flags.