Fix power2unitsString compiler warnings
GCC was warning about unreachable cases, so I corrected the arg type to
match the value being used in the only instance of this function.

I also removed the definition of this function from `RFSpark_wxgui.cpp`,
since it was overriding the static function definition in
`pnlBoardControls.cpp`, where it was actually being used.