Fix a wrong frequency value in comments
At first glance it is unclear where is the error in the comments or in
the code. Until the real intention is not known it looks like an error
in the code where 1.2GHz is used instead of 1.0GHz.

The desired behavior can be found in the documentation (Pg. 18 of the
docs/lms7_api_quick_start_guide.pdf, para: 3.3.2 Device configuration:
"The Rx frequency for both channels is set to 1 GHz while Tx frequency
is set to 1.2 GHz."

Signed-off-by: Gluttton <gluttton@ukr.net>