Improve spurious wakeup handling in USB transfer waiting
The current implementation tries to handle spurious wakeups manually,
but the standard library provides the overloaded version of the
`std::condition_variable::wait_for` that can be used to ignore spurious
awakenings.

Also, the current implementation has a minor logical error: in the case
of false wakeup, the next sleep timeout should be reduced by the time
that was spent on previous waiting.

Signed-off-by: Gluttton <gluttton@ukr.net>