replace mutex of relative time by a custom spinlock
Reading and writing of the relative timestamps pair is a very fast
operation. Moreover, it is expected that contention should be very low.
From the user point of view, it is not necessary to get timestamps too
often. It is enough to get the associated timestamps from time to time
and then and then arbitrary host timestamp value can be easily
calculated by HW clock value. So in most cases the mutex should be free.

To reduce possible performance costs proposed to replace a mutex with a
custom spinlock. The spinlock is built over an `std::atomic_flag` and
satisfies the `Lockable` concept requirements. It means that the Spin
class can be used with standard types (`std::lock_guard` and
`std::unique_lock`).

The proposed solution is controversial. It based on assumption that the
actual mutex implementation is not optimized and has an expensive `lock`
implementation (even if the mutex is free). But in practice, mutex may
be hybrid and may behave like a spinlock. Also OS may provide a cheap
`lock` implementation for free mutex.

Signed-off-by: Gluttton <gluttton@ukr.net>