Fix specification of CMAKE_AUTOSET_INSTALL_RPATH CMake option
The code before this change did not provide any help text to
the option, which made it so CMake defined an option with TRUE
as description, and default value of OFF as per the CMake
documentation:

  https://cmake.org/cmake/help/latest/command/option.html

This can be verified by checking on the CMakeCache.txt:

```
//TRUE
CMAKE_AUTOSET_INSTALL_RPATH:BOOL=OFF
```

This option was introduced in the commit 416510492d4 with an
intent to disable the automatic RPATH logic for Debian packing.

This change merely sanitizes the CMake usage, making it less
confusing for a reader: it seemed that the option was set to
TRUE while the confusingly it was set to OFF. There is no
expected change in behavior with this change.