Fix warning background "sticky" behavior
After a warning error has been logged in the LimeSuiteGUI all the following
messages had an yellow background, which is an unexpected behavior.

Apparently it was caused by the previous fix in the area which made it so the
text color is restored to its default value. Somehow it had an effect on the
background color behavior: as if there is some implicit behavior when neither
of the foreground and background are set in the style when it is set as default
to the message field.

It is hard to tell now if this was an oversight in the previous fix, or is it
something changed in the wxWidgets library. In any case the logging behavior
has been verified once again on both Dark and Light themes on macOS.